test(gc): move slice GC regression test to runtime_roots/string_slice submodule
Keeps runtime_roots.rs (already at 1979 LOC) under the 2000-line file-size gate.
